requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pymysql==1.1.0
cryptography==42.0.0
bcrypt==4.1.2
pytest==7.4.3